Add a module that contains default plotting information.
This module contains the default colormap, norms for different variables, and sets up quantities to use mathsf for latex formatting. I'm unsure whether this is the right spot long term, or if it should live alongside plotting scripts. Either way, the module is not imported automatically, so its impact is minimal if not requested.