Add conversion of MomentInfo to string.
Adding a  __str__ method makes it easy to convert the MomentInfo to a
string for titles and labels. I'm not convinced, however, that as it
exists right now is as universal as I think it is. Also, I'm not sure
this particular mechanism is all that extensible. This might make more
sense as something to do for an individual script and just provide a
few useful pieces.