Refactor multipanel plotting slightly.
Move out a bunch of the code into helper functions. This way when common
plots that share colorbars in rows/columns, or not at all are made, the
tuning of spacing and whatnot can be shared.

It will likely take making quite a few more of these plots using the
helpers to see what's truly useful and where the commonalities are.